+++ core/vpgl/algo/vpgl_fm_compute_7_point.cxx
@@ -51,7 +51,7 @@ vpgl_fm_compute_7_point::compute(
   }
 
   // Construct the design matrix from the point correspondences.
-  vnl_matrix<double> design_matrix(pr_norm.size(),9);
+  vnl_matrix<double> design_matrix(static_cast<unsigned int>(pr_norm.size()), 9);
   for ( unsigned r = 0; r < pr_norm.size(); r++ ) {
     design_matrix(r,0) = pr_norm[r].x()*pl_norm[r].x();
     design_matrix(r,1) = pr_norm[r].y()*pl_norm[r].x();
