+++ core/vpgl/algo/vpgl_triangulate_points.cxx
@@ -24,7 +24,7 @@ double vpgl_triangulate_points::triangulate(
         vgl_point_3d<double> &point_3d)
 {
     constexpr int num_vars = 3;// One var for x, y, z of output 3d point
-    const int num_eqs = 2 * points.size();
+    const int num_eqs = static_cast<const int>(2 * points.size());
 
     // Set up the least-squares solution.
     vnl_matrix<double> A(num_eqs, num_vars, 0.0);
